final: implement asset_data_store using rs_merkle with solidity compatible proofs, solidity verifier contract
