+++ src/sybil/Cargo.toml
@@ -17,17 +17,5 @@ canistergeek_ic_rust = "0.3.1"
 siwe = { git = "https://github.com/domwoe/siwe-rs" }
 time = { version = "0.3.15", default-features = false }
 orally-shared = { path = "../orally-shared" }
-eth_trie = "0.1.0"
-rlp = "0.4.6"
-ethereum-types = "0.11.0"
-primitive-types = "0.12.1"
-keccak-hash = "0.8.0"
-bytes = "1.4.0"
-
 rs_merkle = "1.3.0"
-tiny-keccak = "2.0.2"
-sha3 = "0.10.6"
-
-merkle_light = "0.4.0"
 ethers-rs = "0.2.3"
-crypto = "0.4.0"
