+++ src/sybil/src/main.rs
@@ -36,8 +36,8 @@ mod price_fetcher;
 mod evm_interaction;
 mod exchange_rate_canister;
 mod queries;
-mod asset_data_store_final3;
-// mod asset_data_store_new2;
+// mod asset_data_store_final3;
+mod asset_data_store;
 // mod merkle_tree;
 // mod asset_data_store_final;
